requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2.12
djangorestframework==3.13.1
idna==3.3
praw==7.5.0
prawcore==2.3.0
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
typing_extensions==4.1.1
update-checker==0.18.0
urllib3==1.26.8
websocket-client==1.3.1
